Update RopeString tests

+++ tree/rope_string/rope_node.cc
@@ -4,19 +4,19 @@
 namespace algorithms_archive {
 namespace tree {
 struct RopeNode {
-  int weight;
-  int length;
-  int height;
+  uint64_t weight;
+  uint64_t length;
+  uint64_t height;
   std::string data;
   RopeNode *left;
   RopeNode *right;
 
   RopeNode(std::string data_,
-           int weight_,
-           int length_,
+           uint64_t weight_,
+           uint64_t length_,
            RopeNode *left_ = nullptr,
            RopeNode *right_ = nullptr,
-           int height_ = 1)
+           uint64_t height_ = 1)
       : data{data_},
         weight{weight_},
         length{length_},
@@ -25,11 +25,11 @@ struct RopeNode {
         height{height_} {
   }
   RopeNode(std::string &&data_,
-           int &&weight_,
-           int &&length_,
+           uint64_t &&weight_,
+           uint64_t &&length_,
            RopeNode *&&left_ = nullptr,
            RopeNode *&&right_ = nullptr,
-           int &&height_ = 1)
+           uint64_t &&height_ = 1)
       : data{std::move(data_)},
         weight{std::move(weight_)},
         length{std::move(length_)},
